Support multiple test phases
Although there does seem to be a bug where subsequent phases affect
previous phases. To reproduce, run 3 phases with increasesing request
rates. e.g. 10, 30, 60 all with a duration of 2. After run another test
with only one phase with a load profile of duration 2, req rate of 10.

Then compare the graph of the first phase with the second one you ran.
The first one has a much higher latency.

This happens because the first phase is running while the next phases
are running too. To fix this, you have to wait till each phase completes
before going on to the next one. Might not be a bug depending on how you
look at it :P